requirements.txt
dnspython==2.4.2
aiosmtplib==3.0.1
redis==5.0.1
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.19
aiohttp==3.9.1
python-dotenv==1.0.0
email-validator==2.1.0